[ai] Make the filter size fixed
Instead of a target and an actual filter size, which depends on the
output size, there should be only an actual filter size, on which the
output size depends. The output size is now also always checked in
NetworkLayerBase no matter what kind of a network it is.

Also rename MismatchException to InvalidSizeException and allow it to
accept a RasterSize as a parameter.

[#10]